backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.8.2
pydantic-settings==2.4.0
shapely==2.0.4
simplekml==1.3.6
pyyaml==6.0.2
python-multipart==0.0.9
